+++ src/utils/address.ts
@@ -1,0 +1,6 @@
+export const shortenAddress = (text: string, firstLettersAmount: number, lastLettersAmount: number) => {
+  const arrayFromString = text.split('')
+  const firstLetters = arrayFromString.slice(0, firstLettersAmount).join('')
+  const lastLetters = arrayFromString.slice(arrayFromString.length - 1 - lastLettersAmount).join('')
+  return `${firstLetters}...${lastLetters}`
+}
